+++ src/index.ts
@@ -1,2 +1,2 @@
 export * from "./result.ts";
-export * from "./dataTrail.ts";
+export * from "./trail.ts";